+++ core_client/base/models/v3/config_compress.py
@@ -1,0 +1,15 @@
+from pydantic import BaseModel
+
+
+class ConfigCompress(BaseModel):
+    """
+    {
+        "encoding": ["string"],
+        "mimetypes": ["string"],
+        "min_length": int
+    }
+    """
+
+    encoding: list[str] | None = None
+    mimetypes: list[str] | None = None
+    min_length: int | None = None