+++ core_client/base/models/v3/process_state_progress_io.py
@@ -78,6 +78,9 @@ class ProcessStateProgressIO(BaseModel):
     gop: float | None = None
     drop: int | None = None
     tee: list[ProcessStateProgressIOTee] | None = None
+    level: int | None = None
+    profile: int | None = None
+    sample_fmt: str | None = None
 
     @model_validator(mode="before")
     def remove_empty(cls, values):